Revert "Make game framerate-independent with delta-time scaling"
This reverts commit d947bf43ac4aff42715f089a39cf284f60685986.

+++ src/dudemeyer.h
@@ -68,8 +68,8 @@ class dudemeyer
 		anim_user	 dudemeyer_shooting[8];
 		int32 		 col_detect_x[8];
 		int32 		 col_detect_y[8];
-		float			 ice_delay_x;
-		float			 ice_delay_y;
+		int32			 ice_delay_x;
+		int32			 ice_delay_y;
 		
 	public:
 		anim_source	 dudemeyer_deaths[8];
